+++ CHANGELOG.md
@@ -4,6 +4,25 @@
 
 ### Added
 
+- More CRD documentation ([#354]).
+
+### Changed
+
+- `operator-rs` `0.56.1` -> `0.57.0` ([#354]).
+- Increase resource defaults ([#352]).
+
+### Fixed
+
+- BREAKING: Fixed various issues in the CRD structure. `clusterConfig.credentialsSecret` is now mandatory ([#353]).
+
+[#352]: https://github.com/stackabletech/airflow-operator/pull/352
+[#353]: https://github.com/stackabletech/airflow-operator/pull/353
+[#354]: https://github.com/stackabletech/airflow-operator/pull/354
+
+## [23.11.0] - 2023-11-24
+
+### Added
+
 - [BREAKING] Implement KubernetesExecutor ([#311]).
 - Default stackableVersion to operator version ([#312]).
 - Support PodDisruptionBudgets ([#330]).